LPS-162066 Move normalizeDropdownItems into utils
