+++ modules/apps/knowledge-base/knowledge-base-web/src/main/resources/META-INF/resources/admin/js/components/ActionsDropdown.js
@@ -17,7 +17,7 @@ import {ClayDropDownWithItems} from '@clayui/drop-down';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import normalizeDropdownItems from './normalizeDropdownItems';
+import normalizeDropdownItems from '../utils/normalizeDropdownItems';
 
 export default function ActionsDropdown({actions}) {
 	return actions?.length ? (
