+++ modules/apps/knowledge-base/knowledge-base-web/src/main/resources/META-INF/resources/admin/js/utils/normalizeDropdownItems.js
